[Fix] 확률 수치 조정, 카드 게임 포커스시 스케일 조절, 각종 이름 뜨는 부분 수정, 통발/그물도 잡히면 바로 도감등록 되게 수정

+++ 25_2_fishbattle/Assets/Scripts/SO/Fish/FishSO.cs
@@ -13,7 +13,7 @@ public class FishSO : ScriptableObject
     public int Hp;
     public int AbilityToAct;
     public FishHabitatType HabitatType;
-    public int Probability;
+    public float Probability;
     public int MaxStackSize;
     public bool IsPlayerCard = false;
     public Sprite Icon;
